fix(storage): deduplicate live_cursor_lag_sample DDL (#1753)
The table schema was defined twice identically in cursor.py and
cursor_lag_baseline.py. A direct cross-import between those modules
creates a cycle through daemon/__init__, so the DDL is extracted to
a leaf module polylogue/sources/live/_lag_sample_ddl.py with no
intra-package imports. Both consumers import from there.

Closes #1753
